modified python and css files to meet project requirements

+++ vagrant/catalog/flask_project.py
@@ -1,10 +1,12 @@
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
+from flask import Flask, render_template,\
+    request, redirect, url_for, flash, jsonify
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from tables import Base, Genre, Band, User
 # importing for password security
 from flask import session as login_session
-import random, string
+import random
+import string
 # importing for OAuth
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
@@ -20,8 +22,8 @@ CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Music Application"
 
-#Connect to database
-engine = create_engine('sqlite:///genresofmusicwithusers.db')
+# Connect to database
+engine = create_engine('sqlite:///favoritemusic.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -32,7 +34,8 @@ session = DBSession()
 @app.route('/login')
 def showLogin():
     state = ''.join(
-        random.choice(string.ascii_uppercase + string.digits) for x in range(32))
+        random.choice(string.ascii_uppercase + string.digits)
+        for x in range(32))
     login_session['state'] = state
     # return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
@@ -76,7 +79,6 @@ def gconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
 
-
     # Verify that the access token is used for the intended user.
     gplus_id = credentials.id_token['sub']
     if result['user_id'] != gplus_id:
@@ -84,7 +86,6 @@ def gconnect():
             json.dumps("Token's user ID doesn't match given user ID."), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-
 
     # Verify that the access token is valid for this app.
     if result['issued_to'] != CLIENT_ID:
@@ -97,16 +98,14 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(json.dumps
+                                 ('Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
-
 
     # Store the access token in the session for later use.
     login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
-
 
     # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
@@ -118,7 +117,6 @@ def gconnect():
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
-
 
     # see if user exists, if not, make a new one
     user_id = getUserID(login_session['email'])
@@ -132,7 +130,9 @@ def gconnect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius:' \
+              ' 150px;-webkit-border-radius:' \
+              ' 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
@@ -161,20 +161,21 @@ def getUserID(email):
         return None
 
 
-
-    # DISCONNECT - Revoke a current user's token and reset their login_session
+# DISCONNECT - Revoke a current user's token and reset their login_session
 @app.route('/gdisconnect')
 def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
         print 'Access Token is None'
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     print 'In gdisconnect access token is %s', access_token
     print 'User name is: '
     print login_session['username']
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' \
+          % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
@@ -185,16 +186,21 @@ def gdisconnect():
         del login_session['username']
         del login_session['email']
         del login_session['picture']
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
+
+        response = make_response(json.dumps(
+            'Successfully disconnected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
+        return redirect('/')
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
+        return redirect('/')
 
 
-#JSON APIs to view Genre Information
+# JSON APIs to view Genre Information #DONE
 @app.route('/genres/<int:genre_id>/band/JSON')
 def bandItem(genre_id):
     genre = session.query(Genre).filter_by(id=genre_id).one()
@@ -209,6 +215,7 @@ def bandItemJSON(genre_id, band_id):
     bandItem = session.query(Band).filter_by(id=band_id).one()
     return jsonify(BandItem=bandItem.serialize)
 
+
 @app.route('/genres/JSON')
 def genresJSON():
     genres = session.query(Genre).all()
@@ -220,19 +227,27 @@ def genresJSON():
 @app.route('/genres')
 def homePage():
     genre = session.query(Genre).order_by(Genre.name)
-    return render_template('main.html', genre=genre)
+    if 'username' not in login_session:
+        return render_template('publicgenres.html', genre=genre)
+    else:
+        return render_template('main.html', genre=genre)
 
 
-
-# Show genre information
+# Show bands information in particular genre
 @app.route('/genres/<int:genre_id>/')
 @app.route('/genres/<int:genre_id>/band/')
 def genreItem(genre_id):
     genre = session.query(Genre).filter_by(id=genre_id).one()
+    creator = getUserInfo(genre.user_id)
     items = session.query(Band).filter_by(
         genre_id=genre_id).all()
-    return render_template('band.html', items=items, genre=genre)
-
+    if 'username' not in login_session or \
+            creator.id != login_session['user_id']:
+        return render_template('band.html',
+                               items=items, genre=genre, creator=creator)
+    else:
+        return render_template('band.html',
+                               items=items, genre=genre, creator=creator)
 
 
 # Create new Genre
@@ -242,7 +257,7 @@ def newGenre():
         return redirect('/login')
     if request.method == 'POST':
         newGenre = Genre(
-            name=request.form['name'])
+            name=request.form['name'], user_id=login_session['user_id'])
         session.add(newGenre)
         flash('New Genre %s Successfully Created' % newGenre.name)
         session.commit()
@@ -251,14 +266,15 @@ def newGenre():
         return render_template('newGenres.html')
 
 
-
 # Edit a Genre
 @app.route('/genres/<int:genre_id>/edit/', methods=['GET', 'POST'])
 def editGenre(genre_id):
-    if 'username' not in login_session:
-        return redirect('/login')
     editedGenre = session.query(
         Genre).filter_by(id=genre_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if editedGenre.user_id != login_session['user_id']:
+        return redirect('/')
     if request.method == 'POST':
         if request.form['name']:
             editedGenre.name = request.form['name']
@@ -268,15 +284,19 @@ def editGenre(genre_id):
             'editGenre.html', genre=editedGenre)
 
 
-# Delete a restaurant
+# Delete a Genre
 @app.route('/genres/<int:genre_id>/delete/', methods=['GET', 'POST'])
 def deleteGenre(genre_id):
-    if 'username' not in login_session:
-        return redirect('/login')
     genreToDelete = session.query(
         Genre).filter_by(id=genre_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if genreToDelete.user_id != login_session['user_id']:
+        flash("Not Authorized")
+        return redirect("/")
     if request.method == 'POST':
         session.delete(genreToDelete)
+        flash('%s Successfully Deleted' % genreToDelete.name)
         session.commit()
         return redirect(
             url_for('homePage', genre_id=genre_id))
@@ -285,30 +305,37 @@ def deleteGenre(genre_id):
             'deleteGenre.html', genre=genreToDelete, genre_id=genre_id)
 
 
-
 # Create new band
 @app.route('/genres/<int:genre_id>/band/new', methods=['GET', 'POST'])
 def newGenreItem(genre_id):
     if 'username' not in login_session:
         return redirect('/login')
+    genre = session.query(Genre).filter_by(id=genre_id).one()
+    if login_session['user_id'] != genre.user_id:
+        return redirect('/')
     if request.method == 'POST':
         newItem = Band(name=request.form['name'], description=request.form[
-                           'description'], year=request.form['year'], genre_id=genre_id)
+                           'description'], year=request.form['year'],
+                       genre_id=genre_id, user_id=genre.user_id)
         session.add(newItem)
         session.commit()
+        flash('New Band %s Item Successfully Created' % (newGenreItem))
         return redirect(url_for('genreItem', genre_id=genre_id))
     else:
         return render_template('newgenreitem.html', genre_id=genre_id)
 
 
-
-#Edit current band
+# Edit current band
 @app.route('/genres/<int:genre_id>/band/<int:band_id>/edit',
            methods=['GET', 'POST'])
 def editGenreItem(genre_id, band_id):
     if 'username' not in login_session:
         return redirect('/login')
     editedItem = session.query(Band).filter_by(id=band_id).one()
+    genre = session.query(Genre).filter_by(id=genre_id).one()
+    if login_session['user_id'] != genre.user_id:
+        flash("Not authorized")
+        return redirect('/')
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -318,12 +345,13 @@ def editGenreItem(genre_id, band_id):
             editedItem.price = request.form['year']
         session.add(editedItem)
         session.commit()
+        flash('Band Item Successfully Edited')
         return redirect(url_for('genreItem', genre_id=genre_id))
     else:
 
         return render_template(
-            'editgenreitem.html', genre_id=genre_id, band_id=band_id, item=editedItem)
-
+            'editgenreitem.html', genre_id=genre_id,
+            band_id=band_id, item=editedItem)
 
 
 # Delete a Band
@@ -332,14 +360,17 @@ def editGenreItem(genre_id, band_id):
 def deleteGenreItem(genre_id, band_id):
     if 'username' not in login_session:
         return redirect('/login')
+    genre = session.query(Genre).filter_by(id=genre_id).one()
     itemToDelete = session.query(Band).filter_by(id=band_id).one()
+    if login_session['user_id'] != genre.user_id:
+        return redirect('/')
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
+        flash('Band Item Successfully Deleted')
         return redirect(url_for('genreItem', genre_id=genre_id))
     else:
         return render_template('deletegenreitem.html', item=itemToDelete)
-
 
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
